+++ src/assets/svg/icons/CloseIcon.vue
@@ -1,9 +1,10 @@
 /* Font Awesome Pro 6.3.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license
 (Commercial License) Copyright 2023 Fonticons, Inc. */
 <template>
-	<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 320 512" width="1em" height="1em" fill="currentColor">
+	<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512" width="1em" height="1em" fill="currentColor">
+		<!--! Font Awesome Pro 6.3.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license (Commercial License) Copyright 2023 Fonticons, Inc. -->
 		<path
-			d="M292.6 166.6L315.3 144 270 98.7l-22.6 22.6L158 210.7 68.6 121.4 46 98.7 .7 144l22.6 22.6L112.7 256 23.4 345.4 .7 368 46 413.3l22.6-22.6L158 301.3l89.4 89.4L270 413.3 315.3 368l-22.6-22.6L203.3 256l89.4-89.4z"
+			d="M68.6 57.4L46 34.7 .7 80l22.6 22.6L176.7 256 23.4 409.4 .7 432 46 477.3l22.6-22.6L222 301.3 375.4 454.6 398 477.3 443.3 432l-22.6-22.6L267.3 256 420.6 102.6 443.3 80 398 34.7 375.4 57.4 222 210.7 68.6 57.4z"
 		/>
 	</svg>
 </template>